Kernel: Move Mutex into Locking/
