+++ Kernel/DoubleBuffer.h
@@ -8,7 +8,7 @@
 
 #include <AK/Types.h>
 #include <Kernel/KBuffer.h>
-#include <Kernel/Mutex.h>
+#include <Kernel/Locking/Mutex.h>
 #include <Kernel/Thread.h>
 #include <Kernel/UserOrKernelBuffer.h>
