+++ Kernel/Storage/PATADiskDevice.h
@@ -11,7 +11,7 @@
 #pragma once
 
 #include <Kernel/Interrupts/IRQHandler.h>
-#include <Kernel/Mutex.h>
+#include <Kernel/Locking/Mutex.h>
 #include <Kernel/Storage/StorageDevice.h>
 
 namespace Kernel {
