+++ Kernel/Storage/SATADiskDevice.h
@@ -7,7 +7,7 @@
 #pragma once
 
 #include <Kernel/Interrupts/IRQHandler.h>
-#include <Kernel/Mutex.h>
+#include <Kernel/Locking/Mutex.h>
 #include <Kernel/Storage/AHCIPort.h>
 #include <Kernel/Storage/StorageDevice.h>
