+++ Kernel/Storage/StorageController.h
@@ -12,8 +12,8 @@
 #include <Kernel/Bus/PCI/DeviceController.h>
 #include <Kernel/Devices/Device.h>
 #include <Kernel/IO.h>
+#include <Kernel/Locking/Mutex.h>
 #include <Kernel/Memory/PhysicalPage.h>
-#include <Kernel/Mutex.h>
 #include <Kernel/PhysicalAddress.h>
 #include <Kernel/Random.h>
 #include <Kernel/WaitQueue.h>
